use MC to calculate scale integration

+++ src/weight.h
@@ -14,7 +14,7 @@ struct variable {
   long int CurrVersion;
   int CurrExtMomBin; // current bin of the external momentum
   double CurrTau;    // current external tau
-  int CurrScale;     // Current (Reference) Scale: Index=1, ..., ScaleBinSize
+  double CurrScale;  // Current (Reference) Scale: Index=1, ..., ScaleBinSize
   array<momentum, MaxLoopNum> LoopMom; // all momentum loop variables
   array<double, MaxTauNum> Tau;        // all tau variables
   array<int, MaxLoopNum> LoopSpin;     // all spin variables
